+++ apps/api-service/src/products/wealthanalyzer/space/space.service.ts
@@ -8,7 +8,7 @@ import { DeleteSpaceCommand } from "./commands/impl/delete-space.command"
 import { CreateSpaceCommand } from "./commands/impl/create-space.command"
 import { CreateSpaceRequestDto } from "./dto/request/create-space.request.dto"
 import { UpdateSpaceCommand } from "./commands/impl/update-space.command"
-import { OnEvent } from "@nestjs/event-emitter"
+import { EventEmitter2, OnEvent } from "@nestjs/event-emitter"
 import { EventMap } from "@/shared/constants/event.map"
 import { FindSpaceByNameQuery } from "./queries/impl/find-space-by-name.query"
 import { AssetService } from "../asset/asset.service"
@@ -18,7 +18,8 @@ export class SpaceService {
   constructor(
     private readonly queryBus: QueryBus,
     private readonly commandBus: CommandBus,
-    private readonly assetService: AssetService
+    private readonly assetService: AssetService,
+    private readonly eventEmitter: EventEmitter2
   ) {}
 
   @OnEvent(EventMap.CreateSpace)
@@ -38,20 +39,25 @@ export class SpaceService {
       new FindAllSpaceQuery(userId, searchKeyword)
     )
 
-    const spacesWithValuation = await Promise.all(
+    return await Promise.all(
       spaces.map(async (space) => {
         const valuation = await this.assetService.calculateSpaceValuation(
           userId,
           space._id.toString()
         )
+        const data: { totalUsage: string | number | null | undefined } = (
+          await this.eventEmitter.emitAsync(
+            EventMap.GetAnalyticsTrend,
+            space._id
+          )
+        ).shift()
         return {
           ...(space.toObject?.() ?? space),
           presentValuation: valuation,
+          analyticsTrend: data?.totalUsage,
         }
       })
     )
-
-    return spacesWithValuation
   }
 
   @OnEvent(EventMap.GetSpaceDetailsById)
@@ -65,9 +71,13 @@ export class SpaceService {
         reqUserId,
         space._id.toString()
       )
+      const data: { totalUsage: string | number | null | undefined } = (
+        await this.eventEmitter.emitAsync(EventMap.GetAnalyticsTrend, space._id)
+      ).shift()
       return {
         ...(space.toObject?.() ?? space),
         presentValuation: valuation,
+        analyticsTrend: data?.totalUsage,
       }
     } catch (error) {
       throw new Error(statusMessages.connectionError)