+++ apps/ui-service/src/shared/constants/api-endpoints.ts
@@ -10,7 +10,6 @@ export const endPoints = {
   signOut: `${apiHost}/auth/signout`,
   updateAttribute: `${apiHost}/auth/attribute`,
   // Platform Service
-  analyticsTrends: `${apiHost}/analytics/trends`,
   createCheckoutSession: `${apiHost}/subscription/checkout`,
   subscribe: `${apiHost}/subscription/subscribe`,
   getProductConfig: `${apiHost}/config/product-config`,